+++ core/cli/src/main.rs
@@ -336,7 +336,11 @@ fn main() {
             println!("WASM generated at: {}", wasm_file_path.to_string_lossy());
         }
         if args.generate_v_output {
-            match wasm_to_v(source_fname, wasm_bytes) {
+            match wasm_to_v(
+                source_fname,
+                wasm_bytes,
+                codegen_output.spec_func_indices(),
+            ) {
                 Ok(v_output) => {
                     let v_file_path = output_path.join(format!("{source_fname}.v"));
                     if let Err(e) = fs::create_dir_all(&output_path) {